create  servlets and templates for save new transaction in base

+++ web/src/main/java/com/infoshareacademy/jbusters/dao/NewTransactionDao.java
@@ -25,7 +25,7 @@ public class NewTransactionDao {
         return entityManager.merge(n);
     }
 
-    public void delete(Long id) {
+    public void delete(int id) {
         final NewTransaction n = entityManager.find(NewTransaction.class, id);
         if (n != null) {
             entityManager.remove(n);
